+++ src/Xbackend/xbackend.h
@@ -18,7 +18,7 @@ public:
 	~xmaster();
 	void init(int argc, char* argv[]);
 	void inputhandler_function();
-	int32_t get_focused_viewport_id();
+	int16_t get_focused_viewport_id();
 	uint16_t handle_masterevent(void *event);
 	bool is_system_mode();
 	xcb_connection_t *con;
@@ -26,7 +26,7 @@ public:
 	//uint16_t handle_event(void *event);
 	
 private:
-	viewport *create_viewport_intern(master *masteridd, int32_t ownidd, void *monitor);
+	viewport *create_viewport_intern(master *masteridd, int16_t ownidd, void *monitor);
 	xcb_intern_atom_cookie_t closecookie;
 	xcb_intern_atom_reply_t* closereply;
 	int32_t numbermonitors=0;
