+++ src/testbackend/testbackend.h
@@ -19,11 +19,11 @@ public:
 	void init (int argc, char* argv[]);
 	void inputhandler_function();
 	uint16_t handle_masterevent(void *event);
-	int32_t get_focused_viewport_id();
+	int16_t get_focused_viewport_id();
 	bool is_system_mode();
 
 private:
-	viewport *create_viewport_intern(master *masteridd, int32_t ownidd, void *t);
+	viewport *create_viewport_intern(master *masteridd, int16_t ownidd, void *t);
 };
 
 int testmain(int argc ,char *argv[]);
